Modified getDbl, which now takes a new argument, int *vary. The int to which this points is set to 1 if the double is preceded by "*" and to 0 otherwise.
Problem: Parameters are allocated in "parseSegment". But at that time, we
don't yet know whether "m" is fixed or variable. Without that information,
it isn't clear how to allocate.

Possible solution: Postpone allocation until parseMix. Put a NULL pointer
there to begin with. This will require various modifications, because
sanity checks currently require access to m, and so do the print functions.