Legofit works. Did a successful run with 4 parameters.
Outstanding issues: There is no way to get the GPTree object from the
best-fitting parameter vector, so I just recalculate at the end. This
results in a slightly different answer, because of simulation error.

The DiffEv points are not parallelized. Calculations currently share memory
and would not work in parallel.