Renamed strndx.{c,h} to strint.{c,h}. The change reflects a change in user interface. In the old code, the _get function inserted keys if they were initially missing. The new one requires that key/value pairs be inserted initially. It is an error to "get" a key that is missing. The revised code passes unit tests.
This change also required changes in xdafreader, which works.

The whole point of this was to require that chromosome labels be specified by
command-line argument in tabpat.c. In the old code, problems would have arisen if
the first data file started on chromosome 2, but the second started on chromosome
1. In this case, "2" would have gotten index 0, and "1" would have gotten index 1.
The program would have read to the end of file 1, trying to find the first "1",
and would have ended up with no data at all.

In the new code, chromosome labels must be given on the command line, and in the
order in which they appear in the input files. This fixes the bug just mentioned.