Modified parseChromosomeLbls so that it accepts ranges in reverse order, such as 22-1.
Modified strint.c so that it uses errno to indicate errors rather than
aborting. This makes it possible for tabpat to write more informative
messages before aborting.

Modified tabpat so that it relies on the index generated by StrInt. All
chromosomes that appear in the data must be anticipated in the --chr
argument. I'm not sure what happens to the bootstrap if some chromosomes
listed in --chr don't appear in the data. I think it should work, but more
slowly.

Still getting NaNs in the bootstrap.