Modified PopNode_randomize. It now does repeated calls to PopNode_randomize_r until it generates a feasible set of parameters.
This is intended to fix a problem that arose when I implemented constrained
parameters. In the old code (without constraints), PopNode_randomize_r always
generates a feasible tree. Now, however, constrained parameters may violate
inequality constraints. The result is that some runs fail because in the
initial set of points, all but one are infeasible.