Modified legofit.c and diffev.c. Option --stateIn specifies the name of a state file, which can be read on input. Option --stateOut specifies the state file to be written at the end. This allows Legofit to pick up where it left off. The format of the state file is as follows:
Line 1: two integers, the number of points and the number of free parameters.

Each remaining line describes one point (one set of parameter values). On these
following lines, each field is a floating-point number. The first number is the
cost value. The remaining numbers are the values of the free parameters.