Modified state.c and legofit.c.
When using multiple --stateIn options, it is now possible to use state files
with differing numbers of points, and to request more points (using --ptsPerDim)
than are given in any single --stateIn file. If the input files don't have as
many points as requested, the number of points is revised downward.