Changed signal handling in Legofit.
Legofit now handles 3 signals: SIGINT, SIGTERM, and SIGUSR1. In response to
SIGINT or SIGTERM, legofit waits until the end of the current DE generation,
then exits gracefully, printing all the usual output. In response to SIGUSR1,
it waits until the end of the current DE generation, then writes to stderr
a summary of the current state of the optimizer.