Changed RAFReader_multiNext, which required changes in sitepat.c and mergeraf.c. Not yet tested.
In the old code, RAFReader_multiNext not only advanced all readers to the
next shared position, it also calculated the derived allele frequency,
assuming that the final reader represents the outgroup. With that setup,
it was impossible to use the code in mergedaf.c, which doesn't need the daf
and doesn't define an outgroup.

In the new code, a separate function (RAFReader_findDaf) is used to
calculate the derived allele frequency, and RAFReader_multiNext doesn't do
this job. Thus, mergedaf can use RAFReader_multiNext by itself, but sitepat
must call both functions, one after the other.