Working on clic.c.
Rewrote the function that parses .legofit files. In the new version, parameter
names and values for each input file are pushed onto a separate FIFO stack.
The array of stacks is then used to build a data matrix. This makes it easier
to deal with an initially-unknown number of parameters. Allocation of arrays
can be delayed until after the stacks are populated, and the numbers of files
and of parameters are known.

I also modified make_covar_matrix, so that it takes the data matrix
described above as input, and the covariance matrix is allocated in main.