Fixed bug in xstate.c, which had not been brought up to date with the latest version of state.c.
I've figured out a problem with legofit. It isn't really a bug; it's a usage
mistake that's very easy to make, and in response to which legofit doesn't give
a useful error message.

The problem arises when you use --stateIn arguments to read in files produced
by previous runs of legofit. The format of these files doesn't include a row of
parameter names. Legofit simply assumes that the parameters are as listed in
the .lgo file. If any of the --stateIn arguments were generated using a
different .lgo file, in which the free parameters were different or were listed
in a different order, then legofit will misinterpret the values it reads from
the --stateIn files.

If you're lucky, these misinterpreted values will make no sense, and the
program will abort with a message about an infeasible tree.

The fix is to rewrite state.c, which handles the --stateIn and --stateOut
arguments. We need to change the file format, so that legofit can make sure
that the parameters are consistent. This shouldn't be hard, because  we did
this for the points file that is used by clic.