First draft of re-write of state file code. It compiles. I haven't tried to run it.
The --stateOut option now causes legofit do write its state in a new format,
which is like the old one but contains an additional line. The new line gives
the labels of all free parameter names.

The old --stateIn argument has been replaced by two new ones: --oldStateIn
reads a state file in the old format. --stateIn reads a state file in the
new format.

It is not possible to use --oldStateIn and --stateIn in a single run of
legofit.