Changed state file code again.
The --oldStateIn argument has been eliminated from legofit. Both types of state
file (old and new formats) can be specified with --stateIn, and the two types
can be intermingled in different --stateIn arguments. Legofit checks the
parameter names with new-format state files but now with old-format state
files.

The first line of the new state file format now has 3 fields. The first 2 are
the same as before: the numbers of points and of parameters. The 3rd field
is the character string "new_format". With this setup, the parser can
distinguish old from new format by reading the first line of the file.