Minor changes to bma.c.
Fixed a typo in the documentation.

Merged two for loops into one in the section of code that assigns values to
arrays bepa_file_names and flat_file_names.

Note: the code for parsing command line arguments is fragile. It will need to
change if we ever need an additional command line argument.