Fixed bug bepe.c.
Legofit output files were being parsed with the wrong function: parseSitData
rather than parseLegofit.

I also changed the names of parseSitData and parseLegofit to
StrDblQueue_parseSitData and StrDblQueue_parseLegofit, to make it clear
that these are part of the API of class StrDblQueue.

I switched from strtok_r to strsep, because strtok_r is deprecated.