Shortened the main program in bepe.c.
Changes involve the following:

1. The real data is now "index 0" in all the arrays. This makes it much easier
to let the various data sets take turns "pretending" to be the real data. We
no longer need separate identifiers for the real data.

2. The mean squared difference between StrDblQueue objects is now calculated
within a separate function. This also allows for much shorter code.

The old code is still there but has been commented out with "#if 0".