Changed bepe's output format.
Added a header at the top, giving the name of the program, the date and
time at which it was compiled and run.

The main portion of output is now in two columns, with no comments. 1st column
is bepe, 2nd is the file whose site frequencies we're trying to predict.