strparmap2 and its unit test are working.
This should work just like strparmap. The only difference is that it's based
on the code in parmap.h and parmap.src. These are generic files that implement a
left-leaning red-black tree without specifying which variable within the structure
is the key. That determination is made by preprocessor macro definitions within
strparmap2.h and strparmap2.c. The goal is to have a single code base for strparmap
and addrparmap.