Version 1.76.
In tinyexpr.c, I no longer free the "name" pointer within objects
of type te_variable. This constitutes a memory leak. However, that pointer
sometimes points to string literals, which should not be freed. And the
number of these objects does not grow without limit. (There are only as many
such objects as there are variables defined in the .lgo file.) So it seems
best to live with the memory leak.