Version 1.81.
Changed format of .bepe and .clic files. The output now contains three
columns instead of two. The third column shows the .legofit file that
was used to calculate each value of bepe and clic.

I've also rewritten booma in an effort to allow a check for consistency
of .legofit files. It doesn't currently work, because the old format
of .bepe and .clic files did not provide the .legofit file. Next step
is to teach booma to parse the new-format .bepe and .clic files.