Version 1.82
The new versions of bepe, clic, and booma all work. Bepe and clic now generate
three columns of output rather than two, and booma uses the extra column
to make sure that the .legofit file names in the .bepe and .clic files agree
with those in the .flat files. I haven't yet checked that xbooma is working.