Modified resid.c.
In the new version, it is no longer necessary to provide a list of .legofit
file names. If none are provided, the program prints a table of site
pattern frequencies, with a row for each site pattern and a column for
each data file. It is still possible to collapse several populations into
a single label.