Modified error output in parse.c.
In the new version, the error message is always printed last. It the old
version, date mismatches would trigger a dump of parameter values, which were
printed after the error message. This made the error message hard to find. In
the new version, the parameter dump (if any) is printed before the error
message. Error messages printed by functions (lower down the call stack) may
still preceed the parameter dump.