Rewrote setpart.c and friends.
The new version doesn't allocate memory and fill an array. It is not
object oriented. Instead, the traverseSetPartitions functions visits
all partitions. For each partition, it calls the "visit" function, which is
passed as a function pointer. There is also a void data pointer, which visit
can read and write. This can be used to pass a data structure to control the
behavior of visit and/or to accumulate results.