Revisized Makefile to make things work with Mac OS Catalina.
In the process, I found a bug in MCvisit, within comb.c. Allocation
of cmpl[n - t] fails if n==t. But in this case, the array cmpl isn't
needed anyway, so I just need to move this code inside an "if" block
that will only run if n > t.