Found a bug and am starting to fix it.
The problem is that the old code assumes that coalescent probabilities are
independent and can be multiplied. This fails when migration events follow
coalescent events, because then there may be two sets of migrants (each derived
from a different parental population) that inherit the same coalescent event.
When these sets of migrants join, the current code incorrectly multiplies the
coalescent probability.

The solution, which I have begun to implement, is to keep event/outcome pairs
for coalescent events in the same linked list as those for migration outcomes.
To this end, I have renamed the MigOutcome structure EventLst. The old files
(migoutcome.{c,h} and xmigoutcome.c) are replaced by eventlst.{c,h} and
xeventlst.c.

The unit test xeventlst.c works, but I have not implemented these changes in
the rest of the code.