xidsetset runs without error.
However, I may have introduced a bug. In the old code IdSetSet_add would sum
probabilities when adding identical IdSetSet objects. That isn't possible in
the new code, because the constituent IdSet objects no longer have
probabilities. (These are now stored in a linked list of type EventLst.) I'm
not sure whether identical IdSet objects ever arise in practice. If not, this
code will work fine. If they do, I'll need to add a "multiplicity" field to the
IdSet structure, which will be initialized to 1 and incremented whenever we
need to add identical IdSet objects.