Added '-d' and '--deterministic' to command-line options of legosim.
Moved "#include <stdint.h>" from binary.c to binary.h.

When incrementing rngseed (an unsigned long), the old code first tested
against ULONG_MAX and set the new value to 0 if we were at the max. This
isn't necessary, because wrapping to 0 is the default behavior for unsigned
integers. I replaced that code with a simple addition.