Modified legosim so that it prints probabilities by default, but can also print expected branch lengths in response to arguments -b or --branch_length.
With the deterministic algorithm, these branch lengths may be biased
downward if the argument to -d is > 0. In this case, legosim prints a
warning.

I also fixed a bug in function tfunc, within file patprob. This was dividing
all branch lengths by nreps, which is appropriate for the stochastic
algorithm but not for the deterministic one. This division has been moved into
gptree.c, where it belongs, and no longer affects the deterministic
algorithm.