Fixed additional leaks. I had missed xsegment.c, because it was not listed in TARGETS within test/Makefile. Now it is.
There was a leak in Segment_dup, which caused the "waiting room" arrays
to be allocated twice, losing the first allocation. This is probably
the cause of the memory growth in legofit.

In the previous commit, I also defined mutex variables to control access to the
hash tables used for storing values of binomial coefficients, Stirling2
numbers, and numbers of set partitions. In a multithreaded environment such as
legofit, the old version might have generated errors, and this could explain
the discrepancy I saw last week between the stochastic and deterministic
algorithms, when analyzing complex models.