Version 2.1.0
Add --epsilon argument to legofit.c. This adds DBL_EPSILON to the branch
length associated with each site pattern, eliminating problems that
arise when a site pattern is present in the data but has probability 0
under the model. Now the cost value of such cases will be large but not
infinite.