Wrote and tested longvec.h and longvec.c, which implement an expandable array of long ints.
Used LongInt to fix a problem in sitepat.c. This program previously had
a fixed-size array that only allowed for the possibility of 24
chromosomes. In data with more chromosomes (or scaffolds or contigs) than
this, the program would write outside the array and cause a segmentation
fault.

I suspect this problem still exists in tabpat and maybe simpat.